+++ processor/src/test/fixtures/expected/com/example/cascade_dispose/ParameterizedComponentCascadeDisposeMethodModel_Arez_EventDrivenValue.java
@@ -1,0 +1,65 @@
+package com.example.cascade_dispose;
+
+import arez.Arez;
+import arez.ArezContext;
+import arez.Component;
+import arez.Disposable;
+import arez.component.internal.ComponentKernel;
+import javax.annotation.Nullable;
+import javax.annotation.processing.Generated;
+import org.realityforge.braincheck.Guards;
+
+@Generated("arez.processor.ArezProcessor")
+final class ParameterizedComponentCascadeDisposeMethodModel_Arez_EventDrivenValue<SourceType extends ParameterizedComponentCascadeDisposeMethodModel.EventTarget, ValueType> extends ParameterizedComponentCascadeDisposeMethodModel.EventDrivenValue<SourceType, ValueType> implements Disposable {
+  private static volatile int $$arezi$$_nextId;
+
+  @Nullable
+  private final ComponentKernel $$arezi$$_kernel;
+
+  ParameterizedComponentCascadeDisposeMethodModel_Arez_EventDrivenValue() {
+    super();
+    final ArezContext $$arezv$$_context = Arez.context();
+    final int $$arezv$$_id = ( Arez.areNamesEnabled() || Arez.areRegistriesEnabled() || Arez.areNativeComponentsEnabled() ) ? ++$$arezi$$_nextId : 0;
+    final String $$arezv$$_name = Arez.areNamesEnabled() ? "com_example_cascade_dispose_ParameterizedComponentCascadeDisposeMethodModel_EventDrivenValue." + $$arezv$$_id : null;
+    final Component $$arezv$$_component = Arez.areNativeComponentsEnabled() ? $$arezv$$_context.component( "com_example_cascade_dispose_ParameterizedComponentCascadeDisposeMethodModel_EventDrivenValue", $$arezv$$_id, $$arezv$$_name ) : null;
+    this.$$arezi$$_kernel = new ComponentKernel( Arez.areZonesEnabled() ? $$arezv$$_context : null, Arez.areNamesEnabled() ? $$arezv$$_name : null, $$arezv$$_id, Arez.areNativeComponentsEnabled() ? $$arezv$$_component : null, null, null, null, false, false, false );
+    this.$$arezi$$_kernel.componentConstructed();
+    this.$$arezi$$_kernel.componentReady();
+  }
+
+  private int $$arezi$$_id() {
+    assert null != this.$$arezi$$_kernel;
+    return this.$$arezi$$_kernel.getId();
+  }
+
+  @Override
+  public boolean isDisposed() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.hasBeenInitialized(), () -> "Method named 'isDisposed' invoked on uninitialized component of type 'com_example_cascade_dispose_ParameterizedComponentCascadeDisposeMethodModel_EventDrivenValue'" );
+    }
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.hasBeenConstructed(), () -> "Method named 'isDisposed' invoked on un-constructed component named '" + ( null == this.$$arezi$$_kernel ? "?" : this.$$arezi$$_kernel.getName() ) + "'" );
+    }
+    return this.$$arezi$$_kernel.isDisposed();
+  }
+
+  @Override
+  public void dispose() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.hasBeenInitialized(), () -> "Method named 'dispose' invoked on uninitialized component of type 'com_example_cascade_dispose_ParameterizedComponentCascadeDisposeMethodModel_EventDrivenValue'" );
+    }
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.hasBeenConstructed(), () -> "Method named 'dispose' invoked on un-constructed component named '" + ( null == this.$$arezi$$_kernel ? "?" : this.$$arezi$$_kernel.getName() ) + "'" );
+    }
+    this.$$arezi$$_kernel.dispose();
+  }
+
+  @Override
+  public String toString() {
+    if ( Arez.areNamesEnabled() ) {
+      return "ArezComponent[" + this.$$arezi$$_kernel.getName() + "]";
+    } else {
+      return super.toString();
+    }
+  }
+}